Re-org test_harness files for consistency
